requirements.txt
python-telegram-bot==20.7
PyMuPDF==1.24.9
img2pdf==0.5.1
Pillow==10.1.0
python-dotenv==1.0.0
